ImageCap/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
pillow>=9.0.0
tqdm>=4.62.0
numpy>=1.20.0
pycocoevalcap>=1.2
matplotlib>=3.5.0
scipy>=1.7.0
pandas>=1.3.0
tensorboard>=2.8.0
pyyaml>=6.0
regex>=2022.1.18
sentencepiece>=0.1.96
scikit-learn>=1.0.0
ftfy>=6.1.0
jsonlines>=3.0.0
accelerate>=0.12.0 